rename glue views
