+++ lib/shortcuts/glue-presto-view.js
@@ -12,9 +12,9 @@ const GlueTable = require('./glue-table');
  * @param {String} [options.TableType='VIRTUAL_VIEW'] hard-wired by this shortcut.
 
  */
-class GlueSparkSQLView extends GlueTable {
+class GluePrestoView extends GlueTable {
   constructor(options = {}) {
-    let { SqlVariables = {} } = options;
+    let { Parameters, SqlVariables = {} } = options;
     const {
       Columns,
       DatabaseName,
@@ -29,21 +29,44 @@ class GlueSparkSQLView extends GlueTable {
 
     const TableType = 'VIRTUAL_VIEW';
     const StoredAsSubDirectories = false;
+    const ViewExpandedText = '/* Presto View */';
+    Parameters = Object.assign(
+      { comment: 'Presto View', presto_view: 'true' },
+      Parameters
+    );
     SqlVariables = Object.assign({ DatabaseName }, SqlVariables);
 
+    const columns = Columns.map((c) => ({
+      name: c.Name,
+      type: c.Type.replace(/string/g, 'varchar')
+    }));
+
     const view = {
-      'Fn::Sub': [
-        originalSql,
-        SqlVariables
-      ]
+      'Fn::Base64': {
+        'Fn::Sub': [
+          JSON.stringify({
+            catalog: 'awsdatacatalog',
+            schema: '${DatabaseName}',
+            originalSql,
+            columns
+          }),
+          SqlVariables
+        ]
+      }
+    };
+
+    const ViewOriginalText = {
+      'Fn::Sub': ['/* Presto View: ${view} */', { view }]
     };
 
     super(
       Object.assign(
         {
           TableType,
+          Parameters,
           StoredAsSubDirectories,
-          ViewOriginalText: view
+          ViewOriginalText,
+          ViewExpandedText
         },
         options
       )
@@ -51,4 +74,4 @@ class GlueSparkSQLView extends GlueTable {
   }
 }
 
-module.exports = GlueSparkSQLView;
+module.exports = GluePrestoView;
