+++ lib/shortcuts/glue-spark-view.js
@@ -12,9 +12,9 @@ const GlueTable = require('./glue-table');
  * @param {String} [options.TableType='VIRTUAL_VIEW'] hard-wired by this shortcut.
 
  */
-class GlueAthenaView extends GlueTable {
+class GlueSparkView extends GlueTable {
   constructor(options = {}) {
-    let { Parameters, SqlVariables = {} } = options;
+    let { SqlVariables = {} } = options;
     const {
       Columns,
       DatabaseName,
@@ -29,44 +29,21 @@ class GlueAthenaView extends GlueTable {
 
     const TableType = 'VIRTUAL_VIEW';
     const StoredAsSubDirectories = false;
-    const ViewExpandedText = '/* Presto View */';
-    Parameters = Object.assign(
-      { comment: 'Presto View', presto_view: 'true' },
-      Parameters
-    );
     SqlVariables = Object.assign({ DatabaseName }, SqlVariables);
 
-    const columns = Columns.map((c) => ({
-      name: c.Name,
-      type: c.Type.replace(/string/g, 'varchar')
-    }));
-
     const view = {
-      'Fn::Base64': {
-        'Fn::Sub': [
-          JSON.stringify({
-            catalog: 'awsdatacatalog',
-            schema: '${DatabaseName}',
-            originalSql,
-            columns
-          }),
-          SqlVariables
-        ]
-      }
-    };
-
-    const ViewOriginalText = {
-      'Fn::Sub': ['/* Presto View: ${view} */', { view }]
+      'Fn::Sub': [
+        originalSql,
+        SqlVariables
+      ]
     };
 
     super(
       Object.assign(
         {
           TableType,
-          Parameters,
           StoredAsSubDirectories,
-          ViewOriginalText,
-          ViewExpandedText
+          ViewOriginalText: view
         },
         options
       )
@@ -74,4 +51,4 @@ class GlueAthenaView extends GlueTable {
   }
 }
 
-module.exports = GlueAthenaView;
+module.exports = GlueSparkView;
