+++ lib/shortcuts/index.js
@@ -15,8 +15,8 @@ module.exports = {
   GlueTable: require('./glue-table'),
   GlueJsonTable: require('./glue-json-table'),
   GlueOrcTable: require('./glue-orc-table'),
-  GlueAthenaView: require('./glue-athena-view'),
-  GlueSparkSQLView: require('./glue-sparksql-view'),
+  GluePrestoView: require('./glue-presto-view'),
+  GlueSparkView: require('./glue-spark-view'),
   hookshot: require('./hookshot'),
   LogSubscriptionLambda: require('./log-subscription-lambda')
 };
